Order the code in MainActivity and fix the login infinite loop bug
Should the user not be logged in, he would be prompted with the
FirebaseUI login activity. If the user wants to exit the app by
pressing the back-button, the lifecycle logic would prompt the
login screen indefinitely in an infinite loop. In order to fix
this bug, I override the onActivityResult() callback and handle
there the exit logic according to the result code passed as
parameter to the function.